requirements.txt
asgiref==3.8.1
black==24.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.11
django-filter==24.1
djangorestframework==3.15.1
drf-yasg==1.21.7
Faker==24.4.0
geographiclib==2.0
geopy==2.4.1
idna==3.6
importlib_metadata==7.1.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.3
Markdown==3.6
MarkupSafe==2.1.5
mypy-extensions==1.0.0
numpy==1.26.4
openapi-codec==1.3.2
packaging==24.0
pandas==2.2.1
pathspec==0.12.1
platformdirs==4.2.0
psycopg2==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
simplejson==3.19.2
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
zipp==3.18.1
